Add/Edit Dish updates sold out status
When a dish is a added or the ingredients are updated, the server automatically compares the new ingredients list of the dish against the ingredients we have in stock. If the dish requires any quantity of ingredients greater than what we have in stock, the dish is marked as sold out.